Replaced inheritance with composition for SenitivityBuilder (upgrade response function base)
