+++ applications/FluidDynamicsApplication/custom_response_functions/drag_response_function.h
@@ -14,7 +14,6 @@
 #define KRATOS_DRAG_RESPONSE_FUNCTION_H_INCLUDED
 
 // System includes
-#include <vector>
 #include <string>
 
 // External includes
@@ -25,7 +24,6 @@
 #include "includes/kratos_parameters.h"
 #include "utilities/variable_utils.h"
 #include "response_functions/adjoint_response_function.h"
-#include "solving_strategies/response_functions/response_function_sensitivity_builder_utility.h"
 
 namespace Kratos
 {
@@ -34,172 +32,6 @@ namespace Kratos
 
 ///@name Kratos Classes
 ///@{
-
-class SensitivityBuilder : protected ResponseFunctionSensitivityBuilderUtility
-{
-public:
-    ///@name Type Definitions
-    ///@{
-
-    KRATOS_CLASS_POINTER_DEFINITION(SensitivityBuilder);
-
-    ///@}
-    ///@name Life Cycle
-    ///@{
-
-    /// Constructor.
-    SensitivityBuilder(Parameters& rParameters, ModelPart& rModelPart, AdjointResponseFunction::Pointer pResponseFunction)
-    : mrModelPart(rModelPart), mpResponseFunction(pResponseFunction)
-    {
-        KRATOS_TRY;
-
-        Parameters default_settings(R"(
-        {
-            "sensitivity_model_part_name": "PLEASE_SPECIFY_SENSITIVITY_MODEL_PART",
-            "nodal_sensitivity_variables": ["SHAPE_SENSITIVITY"],
-            "integrate_in_time": true
-        })");
-
-        Parameters custom_settings = rParameters;
-        custom_settings.ValidateAndAssignDefaults(default_settings);
-
-        mSensitivityModelPartName =
-            custom_settings["sensitivity_model_part_name"].GetString();
-
-        Parameters nodal_sensitivity_variables = custom_settings["nodal_sensitivity_variables"];
-        mNodalSensitivityVariables.resize(nodal_sensitivity_variables.size());
-        for (unsigned int i = 0; i < nodal_sensitivity_variables.size(); ++i)
-            mNodalSensitivityVariables[i] = nodal_sensitivity_variables[i].GetString();
-
-        mIntegrateInTime = custom_settings["integrate_in_time"].GetBool();
-
-        KRATOS_CATCH("");
-    }
-
-    ///@}
-    ///@name Operators
-    ///@{
-
-    ///@}
-    ///@name Operations
-    ///@{
-
-    void Initialize()
-    {
-        KRATOS_TRY;
-
-        Check();
-
-        for (const std::string& r_label : mNodalSensitivityVariables)
-            SetNodalSensitivityVariableToZero(r_label, mrModelPart.Nodes());
-
-        VariableUtils().SetNonHistoricalVariable(UPDATE_SENSITIVITIES, false,
-                                                 mrModelPart.Nodes());
-        VariableUtils().SetNonHistoricalVariable(
-            UPDATE_SENSITIVITIES, true,
-            mrModelPart.GetSubModelPart(mSensitivityModelPartName).Nodes());
-
-        KRATOS_CATCH("");
-    }
-
-    void UpdateSensitivities()
-    {
-        KRATOS_TRY;
-
-        double delta_time;
-        if (mIntegrateInTime)
-            delta_time = -mrModelPart.GetProcessInfo()[DELTA_TIME];
-        else
-            delta_time = 1.0;
-        for (const std::string& r_label : mNodalSensitivityVariables)
-            BuildNodalSolutionStepSensitivities(r_label, mrModelPart, delta_time);
-
-        KRATOS_CATCH("");
-    }
-
-    ///@}
-
-protected:
-    ///@name Protected member Variables
-    ///@{
-
-    ///@}
-    ///@name Protected Operators
-    ///@{
-
-    ///@}
-    ///@name Protected Operations
-    ///@{
-
-    void CalculatePartialSensitivity(Variable<array_1d<double, 3>> const& rVariable,
-                                     Element& rElement,
-                                     Matrix const& rSensitivityMatrix,
-                                     Vector& rPartialSensitivity,
-                                     ProcessInfo const& rProcessInfo) const override
-    {
-        mpResponseFunction->CalculatePartialSensitivity(
-            rElement, rVariable, rSensitivityMatrix, rPartialSensitivity, rProcessInfo);
-    }
-
-    ///@}
-
-private:
-    ///@name Member Variables
-    ///@{
-
-    ModelPart& mrModelPart;
-    AdjointResponseFunction::Pointer mpResponseFunction;
-    std::string mSensitivityModelPartName;
-    std::vector<std::string> mNodalSensitivityVariables;
-    bool mIntegrateInTime;
-
-    ///@}
-    ///@name Private Operators
-    ///@{
-
-    ///@}
-    ///@name Private Operations
-    ///@{
-
-    void Check()
-    {
-        KRATOS_TRY;
-
-        if (mrModelPart.HasSubModelPart(mSensitivityModelPartName) == false)
-            KRATOS_ERROR << "No sub model part \"" << mSensitivityModelPartName
-                         << "\"" << std::endl;
-
-        KRATOS_CATCH("");
-    }
-
-    void SetNodalSensitivityVariableToZero(std::string const& rVariableName,
-                                           NodesContainerType& rNodes)
-    {
-        KRATOS_TRY;
-
-        if (KratosComponents<Variable<double>>::Has(rVariableName) == true)
-        {
-            const Variable<double>& r_variable =
-                KratosComponents<Variable<double>>::Get(rVariableName);
-
-            VariableUtils().SetScalarVar(r_variable, r_variable.Zero(), rNodes);
-        }
-        else if (KratosComponents<Variable<array_1d<double, 3>>>::Has(rVariableName) == true)
-        {
-            const Variable<array_1d<double, 3>>& r_variable =
-                KratosComponents<Variable<array_1d<double, 3>>>::Get(rVariableName);
-
-            VariableUtils().SetVectorVar(r_variable, r_variable.Zero(), rNodes);
-        }
-        else
-            KRATOS_ERROR << "Unsupported variable: " << rVariableName << "." << std::endl;
-
-        KRATOS_CATCH("");
-    }
-
-
-    ///@}
-};
 
 /// A response function for drag.
 /**
